+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Schema.org structured data for aggregateRating and reviews
+
 ## [2.2.1] - 2021-07-26
 
 ### Fixed